Fully complete working plotter.  Geometry changes.
I am smart in this stuff now.  I made a lot of changes to how geometry
stuff works.  The plotter that I added actually has free control from
FOVs and the plotting vector.  It's starting to work now.  I am also
working on the rest of the graphics pipeline, which is considerably
easier than making that plotter.  It took me around 16 months to get it
from inspiration to code.